requirements.txt
numpy==1.24.3
opencv-python==4.11.0.86
Pillow>=8.0.0
pyautogui>=0.9.53
python-xlib>=0.31
pynput>=1.7.6
mouse>=0.7.1
pywin32>=300; sys_platform == "win32"
torch==2.1.0+cu118
torchvision==0.16.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
matplotlib>=3.4.0
supervision==0.18.0
scikit-learn>=0.24.0
segment-anything>=1.0
clip @ git+https://github.com/openai/CLIP.git
paddlex==3.1.3
paddleocr==3.1.0
transformers==4.46.1
ultralytics==8.3.70
einops==0.8.0
timm==1.0.19
PyYAML>=5.4.0
openai
boto3>=1.28.57
google-auth<3,>=2
anthropic[bedrock,vertex]>=0.37.1
requests>=2.31.0
wandb>=0.19.9
python-dotenv
flask==3.0.3
dash==3.0.3
huggingface_hub[hf_transfer]>=0.32.0
dash_cytoscape==1.0.2
langgraph 
langchain
langchain-openai
pytesseract
pygments>=2.15.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
hnswlib>=0.7.0
open-clip-torch>=3.0.0